Restructure objects in project structure
