+++ frontend/mietmiez/src/app/objects/internal/Pet.ts
