+++ frontend/mietmiez/src/app/objects/internal/category.ts
